+++ R/deploy.R
@@ -621,29 +621,6 @@ swap_vanity_urls <- function(content_a, content_b) {
   )
 }
 
-#' Swap Vanity URLs
-#'
-#' Swap the vanity URLs of two pieces of content.
-#' This function is deprecated; please use \code{\link{swap_vanity_urls}}.
-#'
-#' @param from A Content object
-#' @param to A Content object
-#'
-#' @returns A list of the new vanity URLs for `from` and `to`
-#'
-#' @family content functions
-#' @export
-swap_vanity_url <- function(from, to) {
-  lifecycle::deprecate_warn("0.6.0", "swap_vanity_url()", "swap_vanity_urls()")
-  res <- swap_vanity_urls(from, to)
-  return(
-    list(
-      from = res[["content_a"]],
-      to = res[["content_b"]]
-    )
-  )
-}
-
 #' Poll Task
 #'
 #' Polls a task, waiting for information about a deployment. If the task has
@@ -694,7 +671,6 @@ poll_task <- function(task, wait = 1, callback = message) {
   }
   task
 }
-
 
 #' Build a Dashboard URL from a Content Item
 #'