+++ NEWS.md
@@ -7,6 +7,8 @@
 - Removed `get_job()`, which was deprecated in 0.6.0. Instead, use `get_jobs()`
   to get a data frame of jobs, use `get_job_list()` to get a list of job
   objects, and use `get_log()` to get the log for a job (#491).
+- Removed `swap_vanity_url()`, which was deprecated in 0.6.0. Use
+  `swap_vanity_urls()` instead.
 
 # connectapi 0.9.0